+++ Changelog.md
@@ -6,3 +6,4 @@
   - Rename modules / types 
   - Change method name: Pauli::index()
   - Add methods: Pauli::min_register_size(), Pauli::num_nontrivial()
+  - Change method name: Pauli:with_ops()